requirements.txt
amqp==5.0.6
asgiref==3.3.4
auditlog3==1.0.1
beautifulsoup4==4.9.3
billiard==3.6.4.0
bs4==0.0.1
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2.4
django-3-jet==1.0.8
django-appconf==1.0.4
django-ckeditor==6.2.0
django-cors-headers==3.7.0
django-currencies==0.10.1
django-image-cropping==1.5.0
django-jazzmin==2.4.9
django-jet==1.0.8
django-js-asset==1.2.2
django-modeltranslation==0.17.3
django-phonenumber-field==5.2.0
django-suit==0.2.25
django-timezone-field==4.2.1
django-utils-six==2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
easy-thumbnails==2.7.1
emoji-country-flag==1.3.0
gunicorn==20.1.0
idna==2.10
install==1.3.4
jsonfield==2.1.1
kombu==5.1.0
Markdown==3.3.4
oauthlib==3.1.1
Pillow==8.3.1
prompt-toolkit==3.0.19
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==2.1.0
python-crontab==2.5.1
python-dateutil==2.8.1
python-gettext==4.0
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sorl-thumbnail==12.7.0
soupsieve==2.2.1
sqlparse==0.4.1
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
